✨ feat: 实现LuaStackAPI

+++ neige_infra/src/state/stack.rs
@@ -5,7 +5,7 @@ pub trait StackApi {
     fn get_top(&self) -> isize;
     fn abs_index(&self, idx: isize) -> isize;
     fn check_stack(&mut self, n: usize) -> bool;
-    fn pop(&mut self, n: u32);
+    fn pop(&mut self, n: isize);
     fn copy(&mut self, from_idx: isize, to_idx: isize);
     fn push_value(&mut self, idx: isize);
     fn replace(&mut self, idx: isize);
